cleanup feature flags, add abort module
